+++ lib/Payplug/Authentication.php
@@ -2,7 +2,7 @@
 namespace Payplug;
 
 /**
- * 
+ *
  */
 class Authentication
 {
@@ -86,7 +86,11 @@ class Authentication
     public static function getPermissionsByLogin($email, $password)
     {
         $keys = self::getKeysByLogin($email, $password);
-        $payplug = Payplug::setSecretKey($keys['httpResponse']['secret_keys']['live']);
+        $payplug = Payplug::init(array(
+            'secretKey' => $keys['httpResponse']['secret_keys']['live'],
+            'apiVersion' => null,
+        ));
+
         $httpClient = new Core\HttpClient($payplug);
         $response = $httpClient->get(Core\APIRoutes::getRoute(Core\APIRoutes::ACCOUNT_RESOURCE));